+++ x/epochs/client/cli/query.go
@@ -63,8 +63,8 @@ func GetCmdEpochsInfos() *cobra.Command {
 				return err
 			}
 
-			// return clientCtx.PrintProto(res)
-			return clientCtx.PrintObjectLegacy(res)
+			return clientCtx.PrintProto(res)
+			//return clientCtx.PrintObjectLegacy(res)
 		},
 	}